Patch files for github actions in main repo
We need to be able to run this repository as a github action as we're
doing the github actions 'as is' and essentially just copying the
workflow within drone.

I've updated the versions of Helm and included the action.yml so that we
can include it in the action over in gomods/athens. Also added the
shellcheck action as I saw that this was an issue as a part of this
repository.

Signed-off-by: Chris Mills <chris.mills@enginegroup.com>